Move models and generator folders; modularize main.py and run
