+++ ml/models/alexnet3d.py
@@ -3,8 +3,10 @@ A simple 3D Convolutional network. It has 5 convolutional
 layers and 3 FC layers, much like Alexnet.
 
 Unfortunately, the architecture described in the paper is
-severely lacking in details. Here I use 2D convolution
-(with depth as channels) instead of 3D. In addition, I assume:
+severely lacking in details. Here, I use full 3D convolutions.
+In addition, I assume:
+- That the depth of the image is kept constant throughout
+  the convolutional layers
 - The filter sizes
 - The strides between convolutional layers
 - The pooling size
@@ -16,13 +18,15 @@ Based on:
 
 from keras.models import Model
 from keras.layers import Input, BatchNormalization, Dense, Flatten
-from keras.layers.convolutional import Conv2D, MaxPooling2D
+from keras.layers.convolutional import Conv3D, MaxPooling3D
+
+from ml.models.model import ModelBuilder
 
 
-class AlexNet2DBuilder(object):
+class AlexNet3DBuilder(ModelBuilder):
 
     @staticmethod
-    def build(input_shape):
+    def build(input_shape, num_classes=2):
         """Create a 3D Convolutional Autoencoder model.
 
         Parameters:
@@ -39,47 +43,43 @@ class AlexNet2DBuilder(object):
         in batch) as input and returns a 5D vector (prediction) as output.
         """
 
-        if len(input_shape) != 3:
+        if len(input_shape) != 4:
             raise ValueError("Input shape should be a tuple "
-                             "(conv_dim1, conv_dim2, conv_dim3)")
+                             "(conv_dim1, conv_dim2, conv_dim3, channels)")
 
         input_img = Input(shape=input_shape, name="cad_input")
 
-        # Conv1 (Output 200 x 200 x 48)
-        x = Conv2D(256, (11, 11), activation='relu',
+        # Conv1 (Output 120 x 120 x 64 x 128)
+        x = Conv3D(128, (7, 7, 7), activation='relu',
                    padding='same')(input_img)
         x = BatchNormalization()(x)
-        x = MaxPooling2D(pool_size=(2, 2), strides=(2, 2))(x)
+        x = MaxPooling3D(pool_size=(2, 2, 2), strides=(2, 2, 2))(x)
 
-        # Conv2 (Output 50 x 50 x 64)
-        x = Conv2D(256, (5, 5), activation='relu', padding='same')(x)
+        # Conv2 (Output 30 x 30 x 16 x 256)
+        x = Conv3D(256, (5, 5, 5), activation='relu', padding='same')(x)
         x = BatchNormalization()(x)
-        x = MaxPooling2D(pool_size=(2, 2), strides=(2, 2))(x)
+        x = MaxPooling3D(pool_size=(2, 2, 2), strides=(2, 2, 2))(x)
 
-        # Conv3 (Output 12 x 12 x 96)
-        x = Conv2D(512, (3, 3), activation='relu',
+        # Conv3 (Output 7 x 7 x 8 x 256)
+        x = Conv3D(256, (3, 3, 3), activation='relu',
                    padding='same')(x)
 
-        # Conv4 (Output 6 x 6 x 128)
-        x = Conv2D(512, (3, 3), activation='relu', strides=(2, 2),
+        # Conv4 (Output 4 x 4 x 4 x 512)
+        x = Conv3D(512, (3, 3, 3), activation='relu', strides=(2, 2, 2),
                    padding='same')(x)
 
-        # Conv5 (Output 3 x 3 x 128)
-        x = Conv2D(1024, (3, 3), activation='relu', strides=(2, 2),
+        # Conv5 (Output 2 x 2 x 2 x 1024)
+        x = Conv3D(1024, (3, 3, 3), activation='relu', strides=(2, 2, 2),
                    padding='same')(x)
 
-        # Flatten
-        x = MaxPooling2D(pool_size=(2, 2), strides=(2, 2))(x)
+        # Pooling and flatten
+        x = MaxPooling3D(pool_size=(2, 2, 2), strides=(2, 2, 2))(x)
         x = Flatten()(x)
 
         # Fully connected layers
         x = Dense(1024, activation='relu', use_bias=True)(x)
         x = Dense(1024, activation='relu', use_bias=True)(x)
-        output_img = Dense(1, activation='softmax', use_bias=True)(x)
+        x = Dense(num_classes, activation='sigmoid', use_bias=True)(x)
 
-        model = Model(inputs=input_img, outputs=output_img)
+        model = Model(inputs=input_img, outputs=x)
         return model
-
-
-m = AlexNet2DBuilder.build((120, 120, 64))
-m.summary()
