+++ ml/models/model.py
@@ -1,0 +1,9 @@
+from abc import abstractmethod
+
+
+class ModelBuilder(object):
+
+    @staticmethod
+    @abstractmethod
+    def build(input_shape, num_classes=2, config={}):
+        pass
