+++ run
@@ -1,8 +1,33 @@
-logfile=$1
+GENERATOR=""
+MODEL=""
+OUTPUT=""
 
-if [ "$#" -ne 1 ]; then
-    echo "Specify a filename. e.g. ./run test.txt"
+while getopts "g:m:o:" option; do
+    case ${option} in
+        g) GENERATOR="$OPTARG";;
+        m) MODEL="$OPTARG";;
+        o) OUTPUT="$OPTARG";;
+    esac
+done
+
+if ! [[ "$GENERATOR" =~ ^(default|mnist|single)$ ]] ; then
+    echo "ERROR: You must provide a correct generator."
+    echo "Available: default | mnist | single"
+    echo "Example usage: ./run -g single -m simple -o result.txt"
     exit 1
 fi
 
-nohup python -u main.py > results/$logfile &
+if ! [[ "$MODEL" =~ ^(alexnet2d|alexnet3d|simple)$ ]] ; then
+    echo "ERROR: You must provide a correct model."
+    echo "Available: alexnet2d | alexnet3d | simple"
+    echo "Example usage: ./run -g single -m simple -o result.txt"
+    exit 1
+fi
+
+if ! test "$OUTPUT" ; then
+    echo "ERROR: You must provide an output file name."
+    echo "Example usage: ./run -g single -m simple -o result.txt"
+    exit 1
+fi
+
+nohup python -u main.py $GENERATOR $MODEL > results/$OUTPUT &
