+++ run_sandbox.py
@@ -1,3 +1,5 @@
-import sandbox.generator_vis
+# import sandbox.generator_vis
 # import sandbox.sandbox2
 # import sandbox.size
+import sys
+print(sys.argv)
